refactor(headless-crawler): ブラウザー操作ライブラリのアーキテクチャを再編成
- lib/core/browser を lib/core/headless-browser と lib/core/page に分離
- ページ別（JobDetail, JobList, JobSearch）にディレクトリを整理
- 機能別（checkers, extractors, validators, navigations, form-fillings）にモジュール分離
- 関心の分離により保守性とテスタビリティを向上
- インポートパスを新しい構造に合わせて更新

BREAKING CHANGE: ブラウザー操作関連のインポートパスが変更されました
