--- apps/headless-crawler/lib/core/browser/index.ts
@@ -1,3 +1,0 @@
-export * from "./assertion";
-export * from "./builder";
-export * from "./interactions";
