--- apps/headless-crawler/lib/core/browser/interactions/extraction/index.ts
@@ -1,1 +1,0 @@
-export * from "./jobDetail";
