--- apps/headless-crawler/lib/core/browser/interactions/form-filling/index.ts
@@ -1,1 +1,0 @@
-export * from "./jobSearch";
