--- apps/headless-crawler/lib/core/browser/interactions/index.ts
@@ -1,4 +1,0 @@
-export * from "./element-action";
-export * from "./extraction";
-export * from "./navigation";
-export * from "./form-filling";
