+++ apps/headless-crawler/lib/core/headless-browser/error.ts
