+++ apps/headless-crawler/lib/core/page/JobList/others/error.ts
@@ -1,0 +1,9 @@
+import { Data } from "effect";
+
+export class ListJobsError extends Data.TaggedError("ListJobsError")<{
+    readonly message: string;
+}> { }
+
+export class IsNextPageEnabledError extends Data.TaggedError(
+    "IsNextPageEnabledError",
+)<{ readonly message: string }> { }
