+++ apps/headless-crawler/lib/core/page/JobSearch/navigations/error.ts
@@ -4,9 +4,6 @@ export class GoToJobListError extends Data.TaggedError("GoToJobListError")<{
   readonly message: string;
 }> {}
 
-export class NextJobListPageError extends Data.TaggedError(
-  "NextJobListPageError",
-)<{ readonly message: string }> {}
 export class JobListPagenationError extends Data.TaggedError(
   "JobListPagenationError",
 )<{ readonly message: string }> {}
